Move `postutils` code into `templating`
Move `postutils` code into `templating` for the
purpose of making all message-parsing go through
Jinja2. This is resource-intensive, but the
future solution is to use flask-caching or some
jinja caching module to update the html output
cache for a thread view, thread list, search list,
etc., whenever a reply or thread is posted.
