+++ tests/test_templating.py
@@ -1,11 +1,14 @@
-from bubblebbs import postutils
+from bubblebbs import templating
 from bubblebbs import models
 
 from . import testutils
 
 
-class TestDatabasePostUtils(testutils.DatabaseTest):
-    """Test post utilities which require a database connection."""
+class TestTemplatesRequiringDb(testutils.DatabaseTest):
+    """Test jinja templating functions which require
+    a database connection.
+    
+    """
 
     def test_reference_links(self):
         """Test the insertion of @2 style post reference links."""
@@ -16,7 +19,7 @@ class TestDatabasePostUtils(testutils.DatabaseTest):
 
         # We feed the raw post text and hope it's correctly parsed
         with self.app.app_context():
-            hopefully_nicely_linked = postutils.reference_links(models.Post, test_links_message, 42)
+            hopefully_nicely_linked = templating.reference_links(models.Post, test_links_message, 42)
 
         # What the post text *should* be after parsing it
         with open('tests/parsing/reference_links_parsed.txt') as f:
@@ -26,7 +29,7 @@ class TestDatabasePostUtils(testutils.DatabaseTest):
 
 
 def test_make_tripcode():
-    assert ('bleh', 'CWj74YsG7iMjTMMxPvhZpA--') == postutils.make_tripcode('bleh#lol')
+    assert ('bleh', 'CWj74YsG7iMjTMMxPvhZpA--') == templating.make_tripcode('bleh#lol')
 
 
 def test_parse_markdown():
@@ -41,7 +44,7 @@ def test_parse_markdown():
 
     # Hopefully this parses correctly!
     # FIXME: using strip() on correct_html_output is bad!
-    hopefully_correct_output = postutils.parse_markdown(markdown_to_parse, unique_slug='cats')
+    hopefully_correct_output = templating.parse_markdown(markdown_to_parse, unique_slug='cats')
     assert hopefully_correct_output == correct_html_output.strip()
 
 
@@ -56,5 +59,5 @@ def test_add_domains_to_link_texts():
         correctly_parsed_message = f.read()
 
     # Hoping this parses correctly using the raw post text!
-    hopefully_correctly_parsed = postutils.add_domains_to_link_texts(message_to_parse)
+    hopefully_correctly_parsed = templating.add_domains_to_link_texts(message_to_parse)
     assert hopefully_correctly_parsed == correctly_parsed_message
